Fix rate limiter
The rate limiter would return wrong values in the HTTP headers after
sliding the window forward because the headers were generated before
adjusting the remaining tokens.

To simplify the rate limiter and make it less error-prone, each visitor
stored in the rate limiter uses a mutex which locks the whole visitor
while operating on its values.